Create README.md
This project is intended to give insights on what types of apps are likely to attract more users. This data will be useful to share with our development team so that they can understand what types of apps to develop so that we can gain the highest number of users. Since we only develop free apps, our main source of revenue are from in-app ads. This is why we are interested in finding out what types of apps attract more users.

Our goal for this project is to determine which apps currently available in the app store and the google play store are attracting users and why.

As a first step, we're going to open both data sets and then slice them into two lists each, so each data set will have a list with all the headers, and a list with the actual data. They will be named subsequently with ios or android to distinguish between the two operating systems.